requirements.txt
Pillow
Django
djangorestframework
djangorestframework-simplejwt
psycopg
drf-yasg